Refactor authentication and entity management, update dependencies, and enhance logging
- Downgraded Spring Boot version from 3.2.1 to 3.1.5 for compatibility.
- Added JwtConfig class for centralized JWT configuration, including secret and expiration properties.
- Updated SecurityConfig to utilize the new JwtConfig and improved role-based access control.
- Refactored various entities and services to change ID types from Long to Integer for consistency.
- Enhanced logging in supermall.log to capture application startup and error details.
- Updated application.yml for clearer JWT configuration and improved logging settings.
- Cleaned up SQL scripts to include a new system_logs table for better tracking of application actions.

These changes improve the overall structure, security, and logging capabilities of the application.

+++ src/main/java/com/supermall/backend/common/security/aspect/PermissionAspect.java
@@ -27,7 +27,7 @@ public class PermissionAspect {
             throw new BusinessException("未登录");
         }
 
-        Long roleId = getRoleIdFromAuthentication(authentication);
+        Integer roleId = getRoleIdFromAuthentication(authentication);
         if (roleId == null) {
             throw new BusinessException("用户角色信息不存在");
         }
@@ -38,7 +38,7 @@ public class PermissionAspect {
         }
     }
 
-    private Long getRoleIdFromAuthentication(Authentication authentication) {
+    private Integer getRoleIdFromAuthentication(Authentication authentication) {
         if (authentication.getPrincipal() instanceof SecurityUser) {
             return ((SecurityUser) authentication.getPrincipal()).getRoleId();
         }
