+++ src/main/java/com/supermall/backend/domain/order/dto/OrderItemResponse.java
@@ -5,9 +5,9 @@ import java.math.BigDecimal;
 
 @Data
 public class OrderItemResponse {
-    private Long id;
-    private Long productId;
-    private Long merchantId;
+    private Integer id;
+    private Integer productId;
+    private Integer merchantId;
     private String productName;
     private String productImage;
     private Integer quantity;
