+++ src/main/java/com/supermall/backend/domain/order/service/ReturnOrderService.java
@@ -9,35 +9,35 @@ public interface ReturnOrderService {
     /**
      * 创建退货订单
      */
-    ReturnOrderResponse createReturnOrder(ReturnOrderRequest request, Long userId);
+    ReturnOrderResponse createReturnOrder(ReturnOrderRequest request, Integer userId);
     
     /**
      * 获取退货订单详情
      */
-    ReturnOrderResponse getReturnOrder(Long returnId, Long userId);
+    ReturnOrderResponse getReturnOrder(Integer returnId, Integer userId);
     
     /**
      * 获取用户的退货订单列表
      */
-    Page<ReturnOrderResponse> getUserReturnOrders(Long userId, int page, int size);
+    Page<ReturnOrderResponse> getUserReturnOrders(Integer userId, int page, int size);
     
     /**
      * 获取商家的退货订单列表
      */
-    Page<ReturnOrderResponse> getMerchantReturnOrders(Long merchantId, int page, int size);
+    Page<ReturnOrderResponse> getMerchantReturnOrders(Integer merchantId, int page, int size);
     
     /**
      * 商家处理退货申请
      */
-    ReturnOrderResponse handleReturnOrder(Long returnId, boolean approved, String handlingNotes, Long operatorId);
+    ReturnOrderResponse handleReturnOrder(Integer returnId, boolean approved, String handlingNotes, Integer operatorId);
     
     /**
      * 确认收到退货
      */
-    ReturnOrderResponse confirmReturn(Long returnId, Long operatorId);
+    ReturnOrderResponse confirmReturn(Integer returnId, Integer operatorId);
     
     /**
      * 确认退款完成
      */
-    ReturnOrderResponse confirmRefund(Long returnId, Long operatorId);
+    ReturnOrderResponse confirmRefund(Integer returnId, Integer operatorId);
 }
